refactor: changed the algokit_transact crate dir structure
