+++ crates/algokit_transact/src/constants.rs
@@ -1,0 +1,6 @@
+
+pub const HASH_BYTES_LENGTH: usize = 32;
+pub const ALGORAND_CHECKSUM_BYTE_LENGTH: usize = 4;
+pub const ALGORAND_ADDRESS_LENGTH: usize = 58;
+pub const ALGORAND_PUBLIC_KEY_BYTE_LENGTH: usize = 32;
+pub type Byte32 = [u8; 32];
