+++ crates/algokit_transact/src/transactions/mod.rs
@@ -1,0 +1,101 @@
+mod asset_transfer;
+mod common;
+mod payment;
+
+pub use asset_transfer::AssetTransferTransactionFields;
+pub use common::{TransactionHeader, TransactionType};
+pub use payment::PayTransactionFields;
+
+use crate::constants::HASH_BYTES_LENGTH;
+use crate::error::AlgoKitTransactError;
+use crate::traits::{AlgorandMsgpack, TransactionId};
+use serde::{Deserialize, Serialize};
+use serde_with::{serde_as, Bytes};
+use std::any::Any;
+
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+#[serde(untagged)]
+pub enum Transaction {
+    Payment(PayTransactionFields),
+    AssetTransfer(AssetTransferTransactionFields),
+}
+
+impl AlgorandMsgpack for Transaction {
+    fn encode(&self) -> Result<Vec<u8>, AlgoKitTransactError> {
+        match self {
+            Transaction::Payment(tx) => tx.encode(),
+            Transaction::AssetTransfer(tx) => tx.encode(),
+        }
+    }
+
+    fn decode(bytes: &[u8]) -> Result<Self, AlgoKitTransactError> {
+        let header = TransactionHeader::decode(bytes)?;
+        match header.transaction_type {
+            TransactionType::Payment => {
+                Ok(Transaction::Payment(PayTransactionFields::decode(bytes)?))
+            }
+            TransactionType::AssetTransfer => Ok(Transaction::AssetTransfer(
+                AssetTransferTransactionFields::decode(bytes)?,
+            )),
+            _ => Err(AlgoKitTransactError::UnknownTransactionType(format!(
+                "{:?}",
+                header.transaction_type
+            ))),
+        }
+    }
+}
+impl TransactionId for Transaction {}
+
+#[serde_as]
+#[derive(Serialize, Deserialize, Debug, PartialEq, Clone)]
+pub struct SignedTransaction {
+    #[serde(rename = "txn")]
+    pub transaction: Transaction,
+
+    #[serde(rename = "sig")]
+    #[serde_as(as = "Bytes")]
+    pub signature: [u8; 64],
+}
+
+impl AlgorandMsgpack for SignedTransaction {
+    const PREFIX: &'static [u8] = b"";
+
+    // Since we provide default values for all transaction fields, serde will not know which
+    // transaction type the bytes actually correspond with. To fix this we need to manually
+    // decode the transaction using Transaction::decode (which does check the type) and
+    // then add it to the decoded struct
+    fn decode(bytes: &[u8]) -> Result<Self, AlgoKitTransactError> {
+        let value: rmpv::Value = rmp_serde::from_slice(bytes)?;
+
+        match value {
+            rmpv::Value::Map(map) => {
+                let txn_value = &map
+                    .iter()
+                    .find(|(k, _)| k.as_str() == Some("txn"))
+                    .unwrap()
+                    .1;
+
+                let mut txn_buf = Vec::new();
+                rmpv::encode::write_value(&mut txn_buf, &txn_value)?;
+
+                let txn = Transaction::decode(&txn_buf)?;
+                let mut stxn: SignedTransaction = rmp_serde::from_slice(bytes)?;
+
+                stxn.transaction = txn;
+
+                return Ok(stxn);
+            }
+            _ => {
+                return Err(AlgoKitTransactError::InputError(format!(
+                    "expected signed transaction to be a map, but got a: {:#?}",
+                    value.type_id()
+                )))
+            }
+        }
+    }
+}
+impl TransactionId for SignedTransaction {
+    fn raw_id(&self) -> Result<[u8; HASH_BYTES_LENGTH], AlgoKitTransactError> {
+        self.transaction.raw_id()
+    }
+}
